+++ src/pages/catalog-product/api/index.ts
@@ -1,0 +1,2 @@
+export { sendingFilterRequest } from "./filter-request";
+export * from "./types";
