+++ src/shared/lib/token/get-token-from-cookie.ts
@@ -1,12 +1,14 @@
+import { TOKEN_NAMES } from "./index";
+
 export function getTokenFromCookie(): string {
   let BEARER_TOKEN = "";
   const arrayCookies = document.cookie.split("; ");
   for (const cookie of arrayCookies) {
     const [name, value] = cookie.split("=");
-    if (name === "user_access_token") {
+    if (name === TOKEN_NAMES.successUserAccess) {
       BEARER_TOKEN = value;
       break;
-    } else if (name === "anonymous_access_token") {
+    } else if (name === TOKEN_NAMES.guestAccess) {
       BEARER_TOKEN = value;
       break;
     }
