+++ src/shared/lib/token/index.ts
@@ -1,2 +1,3 @@
+export { TOKEN_NAMES } from "./token-names";
 export { default as hasLoggedInToken } from "./has-logged-in-token";
 export { getTokenFromCookie } from "./get-token-from-cookie";
